refactor: handlers now contain a new method to return the Router
